% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addobs.r
\name{addobs}
\alias{addobs}
\title{add observations to the data set}
\usage{
addobs(obs)
}
\arguments{
\item{obs}{one of three possible input types:
\itemize{
\item An R data frame with the same columns as the current dataset.
\item A comma-separated string in the following format: "var1=1,var2=2,var3=3" which inputs a single observation.
\item An integer in which case \code{obs} entirely missing observation are added to the dataset
}}
}
\description{
Add observations to the data set, similar in functionality to Stata's append command
}
