% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isloaded.r
\name{assert_loaded}
\alias{assert_loaded}
\title{assert a dataset is loaded in genvar and error otherwise}
\usage{
assert_loaded()
}
\value{
returns NULL, invisibly
}
\description{
assert a dataset is loaded in genvar and error otherwise
}
