% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.r
\name{bigraph}
\alias{bigraph}
\title{command to graph bivariate relationships}
\usage{
bigraph(type, xvars, yvars, xlines = NULL, ylines = NULL,
  title = NULL, xlabel = NULL, ylabel = NULL, xrange, yrange,
  style = NULL, color = NULL, size = NULL, output = "screen",
  resolution = "480x480", file, ...)
}
\arguments{
\item{type}{a quoted list of plot types ("line" for line graphs, "connected" for line graphs with points indicating data points, and "scatter" for graphs with points for the data points).  For example, to plot two lines and one scatter plot on the same graph: "line line scatter".}

\item{xvars}{a varlist in "x1 x2 x3" form giving the variables to plot on the horizontal axis.}

\item{yvars}{a varlist in "y1 y2 y3" form giving the variables to plot on the vertical axis}

\item{...}{other options passed to directly to \code{xyplot} from the \code{lattice} package}
}
\value{
returns NULL, invisibly
}
\description{
bigraph plots bivariate relationships.  it can plot multiple relationships on the same graph.
}
