% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builddata.r
\name{builddata}
\alias{builddata}
\title{creates a dataset of a given number of observations}
\usage{
builddata(n, replace = FALSE)
}
\arguments{
\item{n}{the number of observations to make the new dataset}

\item{replace}{if TRUE, replace a dataset in memory, if FALSE, error if a dataset is already loaded}
}
\value{
returns NULL, invisibly
}
\description{
Creates a dataset of a given number of observations.  Does so by creating a variable called "v1" with all missing values.
}
\examples{
builddata(100, replace=TRUE)
listif()
}
