% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.r
\name{capture}
\alias{capture}
\title{captures an expression, returning 1 if there was an error and zero otherwise}
\usage{
capture(expr, silent = FALSE)
}
\arguments{
\item{expr}{an expression to be evaluated}

\item{silent}{if TRUE,  suppress error messages from printing (default: FALSE)}
}
\value{
0 if the expression successfully ran and 1 otherwise
}
\description{
captures an expression, returning 1 if there was an error and zero otherwise
}
\examples{
capture({log(1)})
capture({log(-1)})
capture({log(x)}) # where x is not an already-created variable
}
