% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.r
\name{clear}
\alias{clear}
\title{clears the dataset in memory}
\usage{
clear()
}
\value{
returns NULL invisibly
}
\description{
removes a dataset from memory, errors if no dataset is loaded
}
\examples{
use(cars, clear=TRUE)
listif()
clear()
listif()
}
