% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.r
\name{describe}
\alias{describe}
\title{lists the names of the variables in the dataset}
\usage{
describe(pattern = NULL)
}
\arguments{
\item{pattern}{an optional regular expression which only returns variable names that match the expression}
}
\description{
lists the names of the variables in the dataset
}
\examples{
use(cars)
describe()
describe("^s")
}
