% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.r
\name{do}
\alias{do}
\title{Executes code on the dataset}
\usage{
do(expr)
}
\arguments{
\item{expr}{an R expression which can use any of the variable names in the current dataset}
}
\description{
Executes an R expression using variables from the dataset.
}
