% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{drop}
\alias{drop}
\title{drops variables or rows from the dataset}
\usage{
drop(x)
}
\arguments{
\item{x}{either a variable list in the format - ~var1+var2+var3 - in which case the listed variables are removed from the dataset, or a condition like: "var1==2" in which case observations that satisfy the condition are removed.}
}
\description{
drops variables or rows from the dataset
}
\examples{
use(cars)
listif()
drop(~speed)
listif()
use(cars)
drop("speed <= 20")
listif()
}
