% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{dropif}
\alias{dropif}
\title{drops rows from the dataset}
\usage{
dropif(x)
}
\arguments{
\item{x}{a condition like (ex: "var1==2") describing the observations that should be removed from the data set.}
}
\description{
drops rows from the dataset
}
\examples{
use(cars)
listif()
dropif("speed <= 20")
listif()
}
