% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.r
\name{estimates_print}
\alias{estimates_print}
\title{display estimation results}
\usage{
estimates_print(name = NULL)
}
\arguments{
\item{name}{name of estimates to be replaced.  If unspecified, print current estimates.}
}
\description{
display estimation results
}
