% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.r
\name{estimates_restore}
\alias{estimates_restore}
\title{restore \code{arata} estimates}
\usage{
estimates_restore(name)
}
\arguments{
\item{name}{name of estimates to be restored}
}
\description{
restore \code{arata} estimates
}
