% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.r
\name{estimates_save}
\alias{estimates_save}
\title{save \code{genvar} estimates}
\usage{
estimates_save(estfile)
}
\arguments{
\item{estfile}{file to save current estimates to.}
}
\value{
returns NULL, invisibly
}
\description{
save \code{genvar} estimates
}
\examples{
use(cars, clear=TRUE)
reg(speed,dist)
fp <- file.path(tempdir(), "myest.rdata")
estimates_save(fp)
clear()
estimates_use(fp)
estimates_get()
}
