% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillin.r
\name{fillin}
\alias{fillin}
\title{Fully rectangularize a dataset}
\usage{
fillin(varlist)
}
\arguments{
\item{varlist}{a variable list in "var1 var2 var3 x*" format where "*" matches zero or more of any character and "?" matches one of any character (an unquoted list will work as well with one variable).  On exit, the data set will contain one observation for every possible interaction of variables with missing values filled in where appropriate.}
}
\value{
returns NULL, invisibly
}
\description{
Make the dataset have one observation for every possible interaction of a list of variables.
}
