% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forvar.r
\name{forvar}
\alias{forvar}
\title{apply a function to each of a list of variables and store the results}
\usage{
forvar(varlist, action, ...)
}
\arguments{
\item{varlist}{a list of variables in the format ~var1+var2+var3+... or as a vector of names like c("var1", "var2", "var3",...).}

\item{action}{an R function to apply to each variable, taking the variable name as its first argument.}

\item{...}{any other options specified are passed to the action function.}
}
\description{
apply a function to each of a list of variables and store the results
}
\examples{
use(cars)

}
