% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.r
\name{gen}
\alias{gen}
\title{generates a new variable that is a transformation of existing variables in the dataset or replaces one}
\usage{
gen(data, form, fun, subset = NULL, replace = FALSE)
}
\arguments{
\item{data}{a data set.  On exit, the dataset contains a newly generated variable or an existing variable is replaced.}

\item{form}{this argument gives the name of the variable to generate, the variables that are used to create the new variable, and an option to generate the variable by a certain set of variables in the following format: newvar~var1+var2+var3|byvar1+byvar2.}

\item{fun}{a function of the variables provided above, the value of which is put in newvar.  For example, if newvar~var1+var2+var3|byvar1+byvar2 were specified and fun=function() sum(var1+var2+var3), then newvar would be the sum of var1, var2, and var3 for each level of (byvar1, byvar2).  In Stata, the same command would be: "by byvar1 byvar2: gen newvar = sum(var1+var2+var3)".}

\item{subset}{only generate values if the condition provided in subset is true.  Make sure to enclose the expression in quotes, like so: subset="female==1 & highschool==1" to generate the values only for women who graduated from highschool.  This option is used like the "if" in Stata.}

\item{replace}{either TRUE or FALSE.  If FALSE (default), the code refuses to alter the variable if the variable already exists.  Otherwise, if replace=TRUE, then the values will be replaced.}
}
\description{
generates a new variable that is a transformation of existing variables in the dataset or replaces one
}
