% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.r
\name{getdata}
\alias{getdata}
\title{exports data frame from genvar environment to R environment}
\usage{
getdata()
}
\value{
the data frame currently in the genvar environment
}
\description{
exports data frame from genvar environment to R environment
}
