% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcode.r
\name{getret}
\alias{getret}
\title{gets a return value from genvar environment}
\usage{
getret(code)
}
\arguments{
\item{code}{the name of the value to get}
}
\value{
returns the value corresponding to \code{code}.
}
\description{
gets a return value from genvar environment
}
\examples{
use(cars, clear=TRUE)
describe()
getret("varlist")
}
