% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headdata.r
\name{headdata}
\alias{headdata}
\title{get first few observations}
\usage{
headdata(num)
}
\arguments{
\item{num}{how many of the first observations to get}
}
\value{
returns the first \code{num} rows of data
}
\description{
get first few observations
}
\examples{
use(cars, clear=TRUE)
headdata(5)
}
