% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isloaded.r
\name{is_loaded}
\alias{is_loaded}
\title{a command to determine whether data is loaded}
\usage{
is_loaded()
}
\value{
returns TRUE if dataset is loaded in genvar and FALSE otherwise
}
\description{
a command to determine whether data is loaded
}
