% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{keep}
\alias{keep}
\title{keeps some variables or rows in the dataset and drops the rest}
\usage{
keep(x)
}
\arguments{
\item{vars}{an optional argument with form, ~var1+var2+var3... Variables var1, var2, var3, ... are kept.  This works like the Stata command: "keep var1 var2 var3 ...".}

\item{rows}{an optional argument which provides a condition that, if true, keeps the row in the dataset (otherwise, the row is dropped).  For example, keep(data,rows="female==1") would keep all women in the data set and drop all men from it.}
}
\description{
keeps some variables or rows in the dataset and drops the rest
}
