% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{keepif}
\alias{keepif}
\title{keeps some rows in the dataset and drops the rest}
\usage{
keepif(x)
}
\arguments{
\item{x}{a condition like: \code{var1==2} in which case observations that satisfy the condition are kept and all others are removed.}
}
\value{
returns NULL, invisibly
}
\description{
keeps some rows in the dataset and drops the rest
}
\examples{
use(cars, clear=TRUE)
keepif(speed <= 20)
listif()
}
