% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.r
\name{L}
\alias{L}
\title{a function to take lags and leads with panel data}
\usage{
L(x, k = 1, ...)
}
\arguments{
\item{x}{variable to lag}

\item{k}{how many lags to take?  If a negative number, leads will be generated.}

\item{...}{other options to pass to \code{plm::lag}, does not need to be specified}
}
\value{
returns lag of the variable as a data frame
}
\description{
a function to take lags and leads with panel data, mostly a wrapper for \code{plm}'s lag function.
}
\examples{
library(plm)
data(Produc)
use(Produc, clear=TRUE)
xtset(year, state)
gen(Lemp, L(emp))
gen(L2emp, L(emp,2))
headdata(10)
}
