% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listif.r
\name{listif}
\alias{listif}
\title{prints the part of the dataset that satisfies certain conditions}
\usage{
listif(cond = NULL, ...)
}
\value{
the part of the dataset that satisfies the condition and contains the specified columns
}
\description{
prints the part of the dataset that satisfies certain conditions
}
