% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.r
\name{logit}
\alias{logit}
\title{estimate a logistic regression}
\usage{
logit(y, x, subset = NULL, weights = NULL, linkfunc = "logit", ...)
}
\arguments{
\item{y}{name of the dependent variable}

\item{x}{names of the independent variables in varlist format, either "x1 x2 x3" or ~x1+x2+X3 format.}

\item{subset}{conditions to run the command only of a subset of the data (analogous to "if" statements in Stata)}

\item{weights}{the name of a variable to use for weights in estimation}

\item{linkfunc}{specify the linking function (logit, by default).  Can set to "probit" to do probit estimation or use \code{probit} (which is equivalent).}

\item{...}{other options to pass to \code{glm}}
}
\value{
b coefficient vector

V covariance matrix of coefficients
}
\description{
estimate a logistic regression
}
