% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.r
\name{mat}
\alias{mat}
\title{create and manipulate matrices}
\usage{
mat(command, ...)
}
\arguments{
\item{command}{matrix subcommand}

\item{...}{arguments to sub-commands}
}
\value{
varies by subcommand
}
\description{
create and manipulate matrices
}
\examples{
todo
}
