% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.r
\name{pred}
\alias{pred}
\title{gets fitted values from a genvar regression object}
\usage{
pred()
}
\value{
returns predictions from model
}
\description{
Gets fitted values from a genvar regression object.  For panel models, this predicts the non-fixed effects part of the regression.
}
\details{
Operates on the loaded estimation object, see \code{estimates_use}.
}
\examples{
use(cars, clear=TRUE)
listif()
reg(dist, speed)
gen(fit, pred())
listif()
}
