% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.r
\name{probit}
\alias{probit}
\title{estimate a probit regression}
\usage{
probit(...)
}
\arguments{
\item{...}{options to pass to \code{logit}}
}
\description{
\code{probit(...)} is equivalent to \code{logit(..., linkfunc="probit")}.
}
