% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produc-data.r
\docType{data}
\name{Produc}
\alias{Produc}
\title{Produc data from the plm package}
\format{An object of class \code{data.frame} with 816 rows and 11 columns.}
\usage{
data(Produc)
}
\description{
An example data set for panel data borrowed from the plm package
}
\keyword{datasets}
