% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.r
\name{rename}
\alias{rename}
\title{renames variables in the dataset}
\usage{
rename(var, newvar)
}
\arguments{
\item{var}{the name of the variable to rename}

\item{newvar}{the new name of the variable}
}
\value{
returns NULL, invisibly
}
\description{
renames variables in the dataset
}
\examples{
use(cars, clear=TRUE)
listif()
rename("speed","velocity")
listif()
}
