% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserve.r
\name{restore}
\alias{restore}
\title{restore a dataset from a previous preserve}
\usage{
restore(envir)
}
\arguments{
\item{envir}{a previous preserve value.}
}
\value{
the preserved data set
}
\description{
restore a dataset from a previous preserve
}
\examples{
require(stats)
data <- cars
p <- preserve(data)
collapse(data,~mean(dist)|speed)
data <- restore(p)
}
