% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savedata.r
\name{savedata}
\alias{savedata}
\title{saves data to a CSV file}
\usage{
savedata(file)
}
\arguments{
\item{file}{a file name to save the current data to}
}
\description{
saves data to a CSV file
}
\examples{
use(Produc)
savedata("Produc.csv")
}
