% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcode.r
\name{setret}
\alias{setret}
\title{sets a return value in genvar environment}
\usage{
setret(code, val)
}
\arguments{
\item{code}{return value name}

\item{val}{value of object}
}
\value{
returns NULL, invisibly
}
\description{
sets a return value in genvar environment
}
\examples{
setret("string", "hello")
getret("string")
}
