% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.varlist.r
\name{subset.varlist}
\alias{subset.varlist}
\title{generate a varlist that is a subset of another}
\usage{
\method{subset}{varlist}(x, vars, ...)
}
\arguments{
\item{x}{a varlist}

\item{vars}{a set of variable names}

\item{...}{currently ignored}
}
\description{
generate a varlist that is a subset of another
}
