% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.r
\name{summarize}
\alias{summarize}
\title{summarize a variable list, giving basic descriptive statistics}
\usage{
summarize(varlist, detail = FALSE)
}
\arguments{
\item{varlist}{a variable list either in "var1 var2 x*" form or, optionally, unquoted.}

\item{detail}{if TRUE, provide a more detailed output for each variable}
}
\value{
returns NULL, invisibly
}
\description{
summarize a variable list, giving basic descriptive statistics
}
\examples{
use(cars, clear=TRUE)
summarize(speed)
summarize("speed dist")
}
