% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchdata.r
\name{switchdata}
\alias{switchdata}
\title{switches data sets among the datasets you have in memory}
\usage{
switchdata(name = "last", clear = FALSE)
}
\arguments{
\item{name}{Name of the dataset to switch to.  The name "last" is reserved for the last dataset loaded (default: last)}

\item{clear}{If TRUE, erase the currently loaded dataset.  If FALSE, make the current dataset the new "last". (default: FALSE)}
}
\value{
returns NULL invisibly
}
\description{
switches data sets among the datasets you have in memory
}
\examples{
use(cars, clear=TRUE)
namedata("cars", clear=TRUE)
library(plm)
data(Produc)
use(Produc, clear=TRUE)
listif()
switchdata("cars")
listif()
switchdata()
listif()
listdata()
}
