% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headdata.r
\name{taildata}
\alias{taildata}
\title{get last few observations}
\usage{
taildata(num)
}
\arguments{
\item{num}{how many of the last few observations to get}
}
\value{
returns last \code{num} rows of data
}
\description{
get last few observations
}
