% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{tostring}
\alias{tostring}
\title{turn a variable of another type into a string variable}
\usage{
tostring(varlist)
}
\arguments{
\item{varlist}{variables to convert, either in the form "var1 var2 var3" or in the form ~var1+var2+var3.}
}
\description{
turn a variable of another type into a string variable
}
