% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varlist.r
\name{varlist}
\alias{varlist}
\title{creates a formula object from a varlist, mostly for internal use.}
\usage{
varlist(x)
}
\arguments{
\item{x}{the varlist to be converted in "var1 var2 var3" format.  Will eventually work with globbing so that "var*" can be used to refer to all variables that begin with var}
}
\value{
a formula object which can be passed to \code{model.frame}
}
\description{
creates a formula object from a varlist, mostly for internal use.
}
