% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtset.r
\name{xtset}
\alias{xtset}
\title{prepares a panel or time series dataset for lag operations}
\usage{
xtset(timevar = NULL, obsvar = NULL)
}
\arguments{
\item{timevar}{the name of the variable to for the time dimension}

\item{obsvar}{the name of the variable to use for the observation dimension}
}
\description{
prepares a panel or time series dataset for lag operations
}
\examples{
use(Produc)
xtset("year", "state")
reg("emp", "unemp", effect="twoway")
reg("emp", "unemp", effect="obs")
reg("emp", "unemp", effect="time")
}
