% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.r
\name{capture}
\alias{capture}
\title{captures an expression, setting \code{getret("error")} to TRUE if there was an error and FALSE otherwise}
\usage{
capture(expr, silent = FALSE)
}
\arguments{
\item{expr}{an expression to be evaluated}

\item{silent}{if TRUE,  suppress error messages from printing (default: FALSE)}
}
\value{
FALSE if the expression successfully ran and TRUE otherwise also sets getret("error") to TRUE or FALSE as well
}
\description{
captures an expression, setting \code{getret("error")} to TRUE if there was an error and FALSE otherwise
}
\examples{
capture({log(1)})
getret("error")
capture({log(-1)})
getret("error")
}
