% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.r
\name{describe}
\alias{describe}
\title{lists the names of the variables in the dataset}
\usage{
describe(pattern)
}
\arguments{
\item{pattern}{an optional regular expression which only returns variable names that match the expression. Can be unquoted if it is just a variable name.}
}
\value{
A vector of names of variables with an attribute called "type" giving the types of the variables.  The class of the object is "varlist".
}
\description{
lists the names of the variables in the dataset
}
\examples{
use(cars, clear=TRUE)
describe()
describe("s*")
describe(speed)
}
