% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{destring}
\alias{destring}
\title{convert a variable with string type into a numeric value}
\usage{
destring(varlist)
}
\arguments{
\item{varlist}{variables to convert, in the form "var1 var2 var3" or, if a single variable, an unquoted variable will work as well (i.e. var1).}
}
\value{
returns NULL, invisibly
}
\description{
convert a variable with string type into a numeric value
}
\examples{
use(cars, clear=TRUE)
tostring(speed)
listif()
describe()
destring(speed)
listif()
describe()
}
