% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{drop}
\alias{drop}
\title{drops variables or rows from the dataset}
\usage{
drop(data, vars = NULL, rows = NULL)
}
\arguments{
\item{data}{a data frame to drop variables from.  On exit, the data frame will not have the variable or rows to drop.}

\item{vars}{an optional argument with form, ~var1+var2+var3... Variables var1, var2, var3, ... are dropped.  This works like the Stata command: "drop var1 var2 var3 ...".}

\item{rows}{an optional argument which provides a condition that, if true, drops the row from the dataset.  For example, drop(data,rows="female==1") would drop all women from the data set.}
}
\description{
drops variables or rows from the dataset
}
