% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{dropif}
\alias{dropif}
\title{drops rows from the dataset}
\usage{
dropif(x)
}
\arguments{
\item{x}{a condition like \code{var1==2} describing the observations that should be removed from the data set.}
}
\value{
returns NULL, invisibly
}
\description{
drops rows from the dataset
}
\examples{
use(cars, clear=TRUE)
listif()
dropif(speed <= 20)
listif()
}
