% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{dropvar}
\alias{dropvar}
\title{drops variables in varlist format from the dataset}
\usage{
dropvar(x)
}
\arguments{
\item{x}{a varlist in "var1 var2 var3" format or unquoted if a single variable}
}
\value{
returns NULL, invisibly
}
\description{
drops variables in varlist format from the dataset
}
\examples{
use(cars, clear=TRUE)
listif()
dropvar(speed)
listif()
}
