% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{dropvar}
\alias{dropvar}
\title{drops variables in varlist format from the dataset}
\usage{
dropvar(x)
}
\arguments{
\item{x}{a varlist either in "var1 var2 var3" format or ~var1+var2+var3 format.}
}
\value{
returns NULL, invisibly
}
\description{
drops variables in varlist format from the dataset
}
\examples{
use(cars, clear=TRUE)
listif()
dropvar("speed")
listif()
use(cars, clear=TRUE)
dropvar(~speed)
listif()
}
