% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.r
\name{estimates_save}
\alias{estimates_save}
\title{save \code{genvar} estimates}
\usage{
estimates_save(file)
}
\arguments{
\item{file}{file to save current estimates to.}
}
\value{
returns NULL, invisibly
}
\description{
save \code{genvar} estimates
}
