% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.r
\name{estimates_store}
\alias{estimates_store}
\title{store \code{genvar} estimates}
\usage{
estimates_store(name)
}
\arguments{
\item{name}{name to use to store current estimates from a \code{genvar} estimation function like \code{reg}, \code{logit}, or \code{probit}.}
}
\value{
returns NULL, invisibly
}
\description{
store \code{genvar} estimates
}
