% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.r
\name{estimates_use}
\alias{estimates_use}
\title{loads \code{genvar} estimates from file}
\usage{
estimates_use(file)
}
\arguments{
\item{file}{file to load estimates from.}
}
\value{
returns NULL, invisibly
}
\description{
loads \code{genvar} estimates from file
}
