% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.r
\name{gen}
\alias{gen}
\title{generates a new variable that is a transformation of existing variables in the dataset or replaces one}
\usage{
gen(form, value, subset = NULL, replace = FALSE)
}
\arguments{
\item{form}{this argument gives the name of the variable to generate, and, optionally, a set of variables to generate by in the following format: newvar~byvar1+byvar2+....}

\item{value}{the transformation of the dataset to replace the "newvar" in option \code{form} with.  For example, form="femalewage~year" and value="sum(wage*female)" to get a variable which has yearly wages if female and a zero if the observation is not female.  In Stata, the same command would be: "by year: egen femalewage = total(wage*female)".}

\item{subset}{only generate values if the condition provided in subset is true.  Make sure to enclose the expression in quotes, like so: subset="female==1 & highschool==1" to generate the values only for women who graduated from highschool.  This option is used like the "if" in Stata.}

\item{replace}{either TRUE or FALSE.  If FALSE (default), the code refuses to alter the variable if the variable already exists.  Otherwise, if replace=TRUE, then the values will be replaced.}
}
\description{
generates a new variable that is a transformation of existing variables in the dataset or replaces one
}
