% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{keep}
\alias{keep}
\title{keeps some variables or rows in the dataset and drops the rest}
\usage{
keep(x)
}
\arguments{
\item{x}{either a variable list in the format - ~var1+var2+var3 - in which case the listed variables are kept in the dataset, the other variables removed, or a condition like: "var1==2" in which case observations that satisfy the condition are kept and all others are removed.}
}
\description{
keeps some variables or rows in the dataset and drops the rest
}
\examples{
use(cars)
listif()
keep(~speed)
listif()
use(cars)
keep("speed <= 20")
listif()
}
