% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.r
\name{keepvar}
\alias{keepvar}
\title{keeps some variables in the dataset and drops the others}
\usage{
keepvar(x)
}
\arguments{
\item{x}{a varlist either of the form "var1 var2 var3" or, if a single variable, it can be unquoted.}
}
\value{
returns NULL, invisibly
}
\description{
keeps some variables in the dataset and drops the others
}
\examples{
use(cars, clear=TRUE)
keepvar(speed)
listif()
}
