% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchdata.r
\name{listdata}
\alias{listdata}
\title{lists datasets currently in memory}
\usage{
listdata()
}
\value{
returns a data frame of the names of the datasets in memory and basic descriptions (number of observations, variables).
}
\description{
lists datasets currently in memory
}
\examples{
use(cars, clear=TRUE)
namedata("cars", clear=TRUE)
library(plm)
data(Produc)
use(Produc, clear=TRUE)
namedata("product", clear=TRUE)
listdata()
namedata("cars_new_name", original="cars", clear=TRUE)
listdata()
}
