% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listif.r
\name{listif}
\alias{listif}
\title{prints the part of the dataset that satisfies certain conditions}
\usage{
listif(cond, vars, ...)
}
\arguments{
\item{cond}{a conditional expression; only observations that satisfy the condition will be returned.}

\item{vars}{a variable list; only variables in the list will be returned.}

\item{...}{other options, currently ignored}
}
\value{
the part of the dataset that satisfies the condition and contains the specified columns
}
\description{
prints the part of the dataset that satisfies certain conditions
}
\examples{
use(cars, clear=TRUE)
listif()
listif(speed <= 20)
}
