% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savedata.r
\name{savedata}
\alias{savedata}
\title{saves data to a CSV or RDS file}
\usage{
savedata(file, rds = FALSE)
}
\arguments{
\item{file}{a file name to save the current data to}

\item{rds}{whether to save the file to an RDS file (default: FALSE)}
}
\value{
returns NULL, invisibly
}
\description{
saves data to a CSV or RDS file
}
\examples{
use(cars, clear=TRUE)
savedata(file.path(tempdir(), "cars.csv"))
savedata(file.path(tempdir(), "cars.rds"), rds=TRUE)
}
