% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.r
\name{tostring}
\alias{tostring}
\title{convert a variable of another type into a string variable}
\usage{
tostring(varlist)
}
\arguments{
\item{varlist}{variables to convert, either in the form "var1 var2 var3" or in the form ~var1+var2+var3.}
}
\value{
returns NULL, invisibly
}
\description{
convert a variable of another type into a string variable
}
\examples{
use(cars, clear=TRUE)
tostring("speed")
listif()
}
