% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.r
\name{use}
\alias{use}
\title{uses a dataset, marking it as the active dataset}
\usage{
use(x, ...)
}
\arguments{
\item{x}{either a data.frame or a csv/dta filename to be imported}
}
\description{
uses a dataset, marking it as the active dataset
}
