% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.r
\name{use}
\alias{use}
\title{uses a dataset, marking it as the active dataset}
\usage{
use(x, ...)
}
\arguments{
\item{x}{usually either a data.frame or a csv/dta filename to be imported.  An R function which returns a data.frame can also be specified.}
}
\description{
uses a dataset, marking it as the active dataset
}
\examples{
Produc
use(Produc)
savedata("Produc.csv")
listif()
dropvar(".*")
}
