% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.r
\name{use}
\alias{use}
\title{uses a dataset, marking it as the active dataset}
\usage{
use(x, clear = FALSE, type = NULL, ...)
}
\arguments{
\item{x}{usually either a data.frame or a csv/dta filename to be imported.  An R function which returns a data.frame can also be specified.}

\item{clear}{if TRUE, erase current data if it already exists. If FALSE, back up data so that it can be switched to later via \code{switchdata} (default: FALSE).}

\item{type}{either "csv" or "dta" for loading csv or dta data set}

\item{...}{other options to pass to \code{read.csv} in case x is a csv file or to \code{read.dta} or \code{read.dta13} depending on the type of file being loaded}
}
\value{
returns NULL invisibly
}
\description{
uses a dataset, marking it as the active dataset
}
\examples{
library(plm)
data(Produc)
use(Produc, clear=TRUE)
listif()
dropvar(".*")
}
